swing-trading-dashboard/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
yfinance>=0.2.28
aiosqlite>=0.19.0
python-multipart>=0.0.6
httpx>=0.27.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pyarrow>=14.0.0
requests>=2.31.0
